Increase LUT bitwidth to 32 Bit
Increase the bitwidth of the LUT to 32 Bit unsigned, up from 16 Bit
unsigned. This fixes numerical issues when using affine quantization in
an unsigned range, where a 16 Bit _signed_ LUT would lead to overflows
in the multiplication results.

+++ test/test_dwconv2d.py
@@ -35,5 +35,5 @@ def test_dwconv2d_forward(N, C, H, W, kernel_size, stride, padding, lut):
     lut = lut.cuda()
 
     native = torch.nn.functional.conv2d(x, w, stride=stride, padding=padding, groups=C)
-    custom = dwconv2d(x, w, lut.short(), stride=stride, padding=padding).float()
+    custom = dwconv2d(x, w, lut, stride=stride, padding=padding).float()
     assert torch.allclose(native, custom)
